Resolve "PwC benchmarks have unsupported scores"

Closes #310

See merge request TIBHannover/orkg/orkg-backend!200

+++ src/main/kotlin/eu/tib/orkg/prototype/paperswithcode/adapters/output/persistence/neo4j/Neo4jEntities.kt
@@ -55,7 +55,7 @@ data class Neo4jBenchmarkUnpacked(
     fun toDatasetSummary() =
         DatasetSummary(
             model,
-            score.toFloat(),
+            score,
             metric,
             paper.resourceId!!,
             paper.label!!,
